mysite/requirements.txt
Django==1.10.5
fcm-django==0.2.7
djangorestframework==3.6.2
django-cors-headers==2.0.2
coreapi==2.3.0
djangorestframework-bulk==0.2.1
